Refactor GetAppliedMigrations for flexibility
We expect some future databases may need more control over the
process to write and read the AppliedMigration data. Instead
of each dialect just proving SQL as a string, we now allow it
to implement a custom function which takes the transaction
as input and should return a slice of AppliedMigration as output.