M1 Mac Support for Testing
The library/mysql container doesn't support arm64/v8,
which caused the tests to not be runnable on M1 Macs.

This commit adds two ways to solve this:

1. TestDB now has a SkippedArchs property, which
allows certain Docker containers to be tagged to
not support certain runtime.GOARCH strings. These
cause a skipped test when used.
2. The MySQL test container was switched from
libary/mysql to mysql/mysql-server (the one produced
directly by the MySQL team), which *does* support
arm64/v8, but only on the latest MySQL 8 version.